Split secret realm into dedicated page
