+++ miniprogram/pages/pve/pve.wxss
@@ -10,52 +10,6 @@ page {
   flex-direction: column;
 }
 
-.tab-bar {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  padding: 0 32rpx;
-  margin-bottom: 16rpx;
-  height: 96rpx;
-  box-sizing: border-box;
-}
-
-.tab-item {
-  flex: 1;
-  text-align: center;
-  font-size: 28rpx;
-  color: #7a81a0;
-  position: relative;
-  padding: 20rpx 0;
-  transition: color 0.2s ease;
-}
-
-.tab-item::after {
-  content: '';
-  position: absolute;
-  left: 50%;
-  bottom: 12rpx;
-  transform: translateX(-50%);
-  width: 48rpx;
-  height: 6rpx;
-  border-radius: 999rpx;
-  background: transparent;
-  transition: background 0.2s ease;
-}
-
-.tab-item--active {
-  color: #3942d5;
-  font-weight: 600;
-}
-
-.tab-item--active::after {
-  background: linear-gradient(90deg, #6a5dff, #9f7bff);
-}
-
-.tab-item--hover {
-  color: #505a85;
-}
-
 .loading {
   padding: 200rpx 0;
   text-align: center;
@@ -64,7 +18,7 @@ page {
 }
 
 .scroll-container {
-  height: calc(100vh - 228rpx);
+  height: calc(100vh - 200rpx);
   padding: 0 32rpx 48rpx;
   box-sizing: border-box;
 }
@@ -88,282 +42,6 @@ page {
   font-size: 32rpx;
   font-weight: 700;
   color: #1b1f36;
-}
-
-.section-actions {
-  display: flex;
-  gap: 16rpx;
-}
-
-.action-btn {
-  border-radius: 999rpx;
-  padding: 0 28rpx;
-}
-
-.level-line {
-  display: flex;
-  justify-content: space-between;
-  color: #59617c;
-  font-size: 24rpx;
-  margin-bottom: 12rpx;
-}
-
-.progress-bar {
-  position: relative;
-  background: #e8ecf6;
-  border-radius: 999rpx;
-  height: 16rpx;
-  overflow: hidden;
-  margin-bottom: 24rpx;
-}
-
-.progress-fill {
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  background: linear-gradient(90deg, #6a5dff, #9f7bff);
-  border-radius: 999rpx;
-}
-
-.attr-grid {
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-  gap: 24rpx;
-}
-
-.attr-item {
-  background: #f7f8ff;
-  border-radius: 20rpx;
-  padding: 24rpx;
-}
-
-.attr-label {
-  font-size: 26rpx;
-  color: #6b7391;
-}
-
-.attr-value {
-  font-size: 40rpx;
-  font-weight: 700;
-  color: #3942d5;
-  margin: 12rpx 0;
-}
-
-.attr-detail {
-  font-size: 22rpx;
-  color: #8c94b0;
-}
-
-.attr-extra {
-  margin-top: 28rpx;
-  display: flex;
-  justify-content: space-between;
-  color: #5a6285;
-  font-size: 24rpx;
-}
-
-.slot-list {
-  display: flex;
-  gap: 24rpx;
-  flex-wrap: wrap;
-}
-
-.slot {
-  flex: 1 1 220rpx;
-  min-width: 220rpx;
-  background: #f8f9fd;
-  border-radius: 24rpx;
-  padding: 24rpx;
-}
-
-.slot-label {
-  font-size: 26rpx;
-  font-weight: 600;
-  color: #38405f;
-  margin-bottom: 16rpx;
-}
-
-.slot-content {
-  display: flex;
-  flex-direction: column;
-  gap: 12rpx;
-}
-
-.slot-name {
-  font-size: 28rpx;
-  font-weight: 600;
-  color: #1f2653;
-  padding-bottom: 8rpx;
-  border-bottom: 4rpx solid #7c89ff;
-}
-
-.slot-meta {
-  font-size: 22rpx;
-  color: #7880a4;
-}
-
-.slot-stats text {
-  display: block;
-  font-size: 22rpx;
-  color: #656d90;
-}
-
-.slot-empty {
-  font-size: 24rpx;
-  color: #a0a6bf;
-}
-
-.inventory-title {
-  margin: 32rpx 0 16rpx;
-  font-size: 26rpx;
-  color: #5a6285;
-  font-weight: 600;
-}
-
-.inventory-empty {
-  margin-bottom: 12rpx;
-}
-
-.inventory-footer {
-  margin-top: 20rpx;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 18rpx 24rpx;
-  border-radius: 20rpx;
-  background: linear-gradient(135deg, rgba(93, 108, 255, 0.12), rgba(152, 199, 255, 0.12));
-  color: #2f3760;
-  transition: transform 0.2s ease, box-shadow 0.2s ease;
-}
-
-.inventory-footer--hover {
-  transform: translateY(-2rpx);
-  box-shadow: 0 12rpx 28rpx rgba(76, 96, 190, 0.18);
-}
-
-.inventory-footer__info {
-  display: flex;
-  align-items: center;
-  gap: 12rpx;
-}
-
-.inventory-footer__icon {
-  font-size: 36rpx;
-}
-
-.inventory-footer__amount {
-  font-size: 30rpx;
-  font-weight: 700;
-}
-
-.inventory-footer__label {
-  font-size: 24rpx;
-  color: #43508a;
-}
-
-.equipment-item {
-  border: 1rpx solid #e3e7f4;
-  border-radius: 24rpx;
-  padding: 24rpx;
-  margin-bottom: 16rpx;
-}
-
-.equipment-head {
-  display: flex;
-  justify-content: space-between;
-  font-size: 28rpx;
-  font-weight: 600;
-}
-
-.equipment-desc {
-  margin: 12rpx 0;
-  color: #606887;
-  font-size: 24rpx;
-}
-
-.equipment-stats text {
-  display: inline-block;
-  margin-right: 12rpx;
-  margin-bottom: 4rpx;
-  font-size: 22rpx;
-  color: #4f5782;
-}
-
-.equipment-actions {
-  margin-top: 16rpx;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.equipment-refine {
-  font-size: 22rpx;
-  color: #7a81a0;
-}
-
-.skill-slots {
-  display: flex;
-  gap: 20rpx;
-  flex-wrap: wrap;
-}
-
-.skill-slot {
-  flex: 1 1 240rpx;
-  background: #f6f7fe;
-  border-radius: 24rpx;
-  padding: 24rpx;
-}
-
-.slot-index {
-  font-size: 24rpx;
-  color: #70789a;
-  margin-bottom: 12rpx;
-}
-
-.skill-detail {
-  display: flex;
-  flex-direction: column;
-  gap: 12rpx;
-}
-
-.skill-name {
-  font-size: 28rpx;
-  font-weight: 700;
-}
-
-.skill-effects text {
-  display: block;
-  font-size: 22rpx;
-  color: #616992;
-}
-
-.skill-item {
-  border: 1rpx solid #e0e5f5;
-  border-radius: 24rpx;
-  padding: 24rpx;
-  margin-bottom: 16rpx;
-}
-
-.skill-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 12rpx;
-}
-
-.skill-desc {
-  color: #5f678a;
-  font-size: 24rpx;
-  margin-bottom: 12rpx;
-}
-
-.skill-actions {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  font-size: 22rpx;
-  color: #7a81a0;
 }
 
 .enemy-card {
@@ -464,10 +142,6 @@ page {
   font-size: 22rpx;
   color: #70779a;
   margin-top: 6rpx;
-}
-
-.history-header {
-  margin-top: 20rpx;
 }
 
 .empty-card {
