+++ miniprogram/pages/role/index.json
@@ -1,0 +1,5 @@
+{
+  "usingComponents": {
+    "custom-nav": "/components/custom-nav/custom-nav"
+  }
+}
